requirements.txt
google-cloud-speech==2.27.0
uvicorn==0.30.5
fastapi==0.112.0
python-multipart==0.0.9
